Remove IsExplored property from Cell and Map
Now thatwe can properly inherit from these classes, it should be up to the developer to determine which additional properties should be a part of cells. Transparency and Walkability are necessary for FOV and Pathfinding. Explored status was only for the demo game and shouldn't actually be part of the RogueSharp library